Fix (tests) import helpers.js Move helpers.js to utils
