+++ src/components/button-group/button-group.spec.js
@@ -1,4 +1,4 @@
-import { loadFixture, testVM } from "../helpers"
+import { loadFixture, testVM } from "../../utils/helpers";
 
 describe("button-group", async () => {
     beforeEach(loadFixture("button-group"))
