+++ src/components/carousel/carousel.spec.js
@@ -1,4 +1,4 @@
-import {loadFixture, testVM} from '../helpers';
+import {loadFixture, testVM} from "../../utils/helpers";
 
 describe('carousel', async() => {
     beforeEach(loadFixture('carousel'));
