+++ src/components/container/container.spec.js
@@ -1,4 +1,4 @@
-import { loadFixture, testVM } from "../helpers";
+import { loadFixture, testVM } from "../../utils/helpers";
 
 describe("form-row", async () => {
     beforeEach(loadFixture("container"));
