+++ src/components/form/form-feedback.spec.js
@@ -1,4 +1,4 @@
-import { loadFixture, testVM } from "../helpers";
+import { loadFixture, testVM } from "../../utils/helpers";
 
 describe("form-feedback", async () => {
     beforeEach(loadFixture("form-feedback"));
