+++ src/components/form/form-text.spec.js
@@ -1,4 +1,4 @@
-import { loadFixture, testVM } from "../helpers";
+import { loadFixture, testVM } from "../../utils/helpers";
 
 describe("form-text", async () => {
     beforeEach(loadFixture("form-text"));
