+++ src/components/form/form.spec.js
@@ -1,4 +1,4 @@
-import { loadFixture, testVM } from "../helpers";
+import { loadFixture, testVM } from "../../utils/helpers";
 
 describe("form", async () => {
     beforeEach(loadFixture("form"));
