+++ src/components/index.js
@@ -1,74 +1,74 @@
-import bAlert from "./alert/alert.vue"
-import bBadge from "./badge/badge"
-import bBreadcrumb from "./breadcrumb/breadcrumb"
-import bBreadcrumbItem from "./breadcrumb/breadcrumb-item"
-import bBreadcrumbLink from "./breadcrumb/breadcrumb-link"
-import bButton from "./button/button"
-import bButtonClose from "./button/button-close"
-import bButtonToolbar from "./button-toolbar/button-toolbar.vue"
-import bButtonGroup from "./button-group/button-group"
-import bInputGroup from "./input-group/input-group.vue"
-import bInputGroupAddon from "./input-group/input-group-addon"
-import bInputGroupButton from "./input-group/input-group-button"
-import bCard from "./card/card"
-import bCardBody from "./card/card-body"
-import bCardHeader from "./card/card-header"
-import bCardFooter from "./card/card-footer"
-import bCardGroup from "./card/card-group"
-import bCardImg from "./card/card-img"
-import bCarousel from "./carousel/carousel.vue"
-import bCarouselSlide from "./carousel/carousel-slide.vue"
-import bCol from "./layout/col"
-import bCollapse from "./collapse/collapse.vue"
-import bContainer from "./container/container"
-import bDropdown from "./dropdown/dropdown.vue"
-import bDropdownItem from "./dropdown/dropdown-item"
-import bDropdownItemButton from "./dropdown/dropdown-item-button"
-import bDropdownDivider from "./dropdown/dropdown-divider"
-import bDropdownHeader from "./dropdown/dropdown-header"
-import bEmbed from "./embed/embed"
-import bForm from "./form/form"
-import bFormRow from "./form/form-row"
-import bFormText from "./form/form-text"
-import bFormFeedback from "./form/form-feedback"
-import bFormGroup from "./form-group/form-group.vue"
-import bFormCheckboxGroup from "./form-checkbox/form-checkbox-group.vue"
-import bFormCheckbox from "./form-checkbox/form-checkbox.vue"
-import bFormRadioGroup from "./form-radio/form-radio-group.vue"
-import bFormRadio from "./form-radio/form-radio.vue"
-import bFormInput from "./form-input/form-input.vue"
-import bFormTextarea from "./form-textarea/form-textarea.vue"
-import bFormFile from "./form-file/form-file.vue"
-import bFormSelect from "./form-select/form-select.vue"
-import bImg from "./img/img"
-import bImgLazy from "./img/img-lazy.vue"
-import bJumbotron from "./jumbotron/jumbotron"
-import bLink from "./link/link"
-import bListGroup from "./list-group/list-group"
-import bListGroupItem from "./list-group/list-group-item"
-import bMedia from "./media/media"
-import bMediaBody from "./media/media-body"
-import bMediaAside from "./media/media-aside"
-import bModal from "./modal/modal.vue"
-import bNav from "./nav/nav"
-import bNavItem from "./nav/nav-item"
-import bNavItemDropdown from "./nav/nav-item-dropdown.vue"
-import bNavToggle from "./nav/nav-toggle.vue"
-import bNavbar from "./navbar/navbar"
-import bNavbarNav from "./navbar/navbar-nav"
-import bNavbarBrand from "./navbar/navbar-brand"
-import bNavText from "./nav/nav-text"
-import bNavForm from "./nav/nav-form"
-import bRow from "./layout/row"
-import bPagination from "./pagination/pagination.vue"
-import bPaginationNav from "./pagination/pagination-nav.vue"
-import bPopover from "./popover/popover.vue"
-import bProgressBar from "./progress/progress-bar.vue"
-import bProgress from "./progress/progress.vue"
-import bTable from "./table/table.vue"
-import bTabs from "./tab/tabs.vue"
-import bTab from "./tab/tab.vue"
-import bTooltip from "./tooltip/tooltip.vue"
+import bAlert from './alert/alert.vue';
+import bBadge from './badge/badge';
+import bBreadcrumb from './breadcrumb/breadcrumb';
+import bBreadcrumbItem from './breadcrumb/breadcrumb-item';
+import bBreadcrumbLink from './breadcrumb/breadcrumb-link';
+import bButton from './button/button';
+import bButtonClose from './button/button-close';
+import bButtonToolbar from './button-toolbar/button-toolbar.vue';
+import bButtonGroup from './button-group/button-group';
+import bInputGroup from './input-group/input-group.vue';
+import bInputGroupAddon from './input-group/input-group-addon';
+import bInputGroupButton from './input-group/input-group-button';
+import bCard from './card/card';
+import bCardBody from './card/card-body';
+import bCardHeader from './card/card-header';
+import bCardFooter from './card/card-footer';
+import bCardGroup from './card/card-group';
+import bCardImg from './card/card-img';
+import bCarousel from './carousel/carousel.vue';
+import bCarouselSlide from './carousel/carousel-slide.vue';
+import bCol from './layout/col';
+import bCollapse from './collapse/collapse.vue';
+import bContainer from './container/container';
+import bDropdown from './dropdown/dropdown.vue';
+import bDropdownItem from './dropdown/dropdown-item';
+import bDropdownItemButton from './dropdown/dropdown-item-button';
+import bDropdownDivider from './dropdown/dropdown-divider';
+import bDropdownHeader from './dropdown/dropdown-header';
+import bEmbed from './embed/embed';
+import bForm from './form/form';
+import bFormRow from './form/form-row';
+import bFormText from './form/form-text';
+import bFormFeedback from './form/form-feedback';
+import bFormGroup from './form-group/form-group.vue';
+import bFormCheckboxGroup from './form-checkbox/form-checkbox-group.vue';
+import bFormCheckbox from './form-checkbox/form-checkbox.vue';
+import bFormRadioGroup from './form-radio/form-radio-group.vue';
+import bFormRadio from './form-radio/form-radio.vue';
+import bFormInput from './form-input/form-input.vue';
+import bFormTextarea from './form-textarea/form-textarea.vue';
+import bFormFile from './form-file/form-file.vue';
+import bFormSelect from './form-select/form-select.vue';
+import bImg from './img/img';
+import bImgLazy from './img/img-lazy.vue';
+import bJumbotron from './jumbotron/jumbotron';
+import bLink from './link/link';
+import bListGroup from './list-group/list-group';
+import bListGroupItem from './list-group/list-group-item';
+import bMedia from './media/media';
+import bMediaBody from './media/media-body';
+import bMediaAside from './media/media-aside';
+import bModal from './modal/modal.vue';
+import bNav from './nav/nav';
+import bNavItem from './nav/nav-item';
+import bNavItemDropdown from './nav/nav-item-dropdown.vue';
+import bNavToggle from './nav/nav-toggle.vue';
+import bNavbar from './navbar/navbar';
+import bNavbarNav from './navbar/navbar-nav';
+import bNavbarBrand from './navbar/navbar-brand';
+import bNavText from './nav/nav-text';
+import bNavForm from './nav/nav-form';
+import bRow from './layout/row';
+import bPagination from './pagination/pagination.vue';
+import bPaginationNav from './pagination/pagination-nav.vue';
+import bPopover from './popover/popover.vue';
+import bProgressBar from './progress/progress-bar.vue';
+import bProgress from './progress/progress.vue';
+import bTable from './table/table.vue';
+import bTabs from './tab/tabs.vue';
+import bTab from './tab/tab.vue';
+import bTooltip from './tooltip/tooltip.vue';
 
 export {
     bAlert,
@@ -168,4 +168,4 @@ export {
     bCarousel,
     bCollapse,
     bLink
-}
+};
