+++ src/components/link/link.spec.js
@@ -1,4 +1,4 @@
-import { loadFixture, testVM } from "../helpers";
+import { loadFixture, testVM } from "../../utils/helpers";
 
 describe("link", async () => {
     beforeEach(loadFixture("link"));
